aihub_chat/requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0.0
openai>=1.12.0
Pillow>=10.0.0
pycryptodome>=3.18.0
xmltodict>=0.13.0
pydub>=0.25.1

